% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_prevalence.R
\name{get_ons_smoking}
\alias{get_ons_smoking}
\title{Get data from ONS on England smoking prevalence}
\usage{
get_ons_smoking()
}
\value{

}
\description{
The data from ONS is in wide form and the spreadsheet has many tabs
This data extraction relates to the data Table 1. Proportion of cigarette smokers,
by sex and age, Great Britain, 1974 to 2019.
}
