% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wide.R
\name{get_smoking_wide}
\alias{get_smoking_wide}
\title{Get data from ONS on England smoking prevalence}
\usage{
get_smoking_wide(clean = TRUE)
}
\arguments{
\item{clean}{Logical, default is to remove blank rows and columns but if the
original is required then \code{clean = FALSE}}
}
\value{
data frame
}
\description{
The data from ONS is in wide form and the spreadsheet has many
tabs. This data extraction relates to the data Table 1. Proportion of
cigarette smokers, by sex and age, Great Britain, 1974 to 2019.
}
