% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_prevalence.R
\name{tidy_ons_smoking}
\alias{tidy_ons_smoking}
\title{Tidy data from ONS on England smoking prevalence}
\usage{
tidy_ons_smoking()
}
\value{
data frame
}
\description{
Because the column names are split between two rows in this dataset,
this function tidies the columns names to merge these data so it reads "Gender Age Band".
The dataset remains in wide form.
}
