% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_long.R
\name{tidy_smoking_long}
\alias{tidy_smoking_long}
\title{Tidy ONS data into long form}
\usage{
tidy_smoking_long(add_year = TRUE, rounding = TRUE, separate_cols = TRUE)
}
\arguments{
\item{add_year}{Logical, add a column with a real date which is the beginning
of the year. Default is to add this column}

\item{rounding}{Logical, rounding the population values to whole numbers to
reduce spurious accuracy. Default is to round to whole number}

\item{separate_cols}{Logical, separates the previously merged columns of
gender and age to separate columns in long form. Default is to separate}
}
\value{
Data frame
}
\description{
Tidy ONS data into long form
}
